mgr: move GIL helpers to Gil.{h,cc}
(cherry picked from commit 88db7b190a974d70c08a1951b06b7fc9957b1b60)
